requirements.txt
fastapi
uvicorn[standard]
duckduckgo-search
ollama
pywebview
python-multipart
pypdf
httpx
beautifulsoup4
python-docx
